Removing tangling comma
Hi Pascal
Thank you for merging! I just tried to add the library to the platformio registry but there seems to be a problem. The tool complains about an invalid json file. The JSON lint tool pointed out that the tangling comma at the end of the email address attribute might be the reason, so I am removing it. If that doesn't satisfy the platformio tool I will contact their team for advice...
Cheers,
Dani